ENG-4099 Move Okta SAML-related functions to okta/aws.ts
Move common Okta SAML-related functions to okta/aws.ts. This is
in preparation for adding a new `p0 aws assume permission-set`
command.
